+++ src/dclassql/__init__.py
@@ -3,6 +3,7 @@ from .db_pool import BaseDBPool, save_local
 from .model_inspector import DataSourceConfig
 from .push import db_push
 from .runtime.backends.lazy import eager
+from .runtime.backends.relation_view import LazyLookupKey, LazyRelationView
 from .runtime.sql_recorder import record_sql
 from .unwarp import unwarp, unwarp_or, unwarp_or_raise
 
@@ -17,5 +18,7 @@ __all__ = [
     'BaseDBPool',
     'save_local',
     'DataSourceConfig',
+    'LazyLookupKey',
+    'LazyRelationView',
     'record_sql',
 ]